requirements.txt
Flask==3.1.2
nltk==3.8.1
pdfplumber==0.11.7
openai==1.107.1
transformers>=4.30.0
torch>=2.0.0
python-dotenv==1.0.0
